requirements.txt
argparse
pymodbus
pysunspec2
pyserial
paho-mqtt